+++ cbits/fontmanager.h
@@ -39,6 +39,8 @@ FMcontext* fmInit(float dpr);
 
 int fmCreateFont(FMcontext* ctx, const char* name, const char* filename);
 
+int fmCreateFontMem(FMcontext* ctx, const char* name, unsigned char* data, int dataSize);
+
 void fmSetScale(FMcontext* ctx, float scale);
 
 void fmFontFace(FMcontext* ctx, const char* font);