+++ src/Monomer/Graphics/Types.hs
@@ -45,11 +45,16 @@ data Color = Color {
 instance Default Color where
   def = Color 255 255 255 1.0
 
--- | The definition of a font.
-data FontDef = FontDef {
-  _fntName :: !Text,  -- ^ The logic name. Will be used when defining styles.
-  _fntPath :: !Text   -- ^ The path in the filesystem.
-} deriving (Eq, Show, Generic)
+data FontDef
+  = FontDefFile
+    { _fntFontName :: !Text  -- ^ The logic name. Will be used when defining styles.
+    , _fntFontPath :: !Text  -- ^ The path in the filesystem.
+    }
+  | FontDefMem
+    { _fntFontName :: !Text         -- ^ The logic name. Will be used when defining styles.
+    , _fntFontBytes :: !ByteString  -- ^ The bytes of the loaded font.
+    }
+  deriving (Eq, Show, Generic)
 
 -- | The name of a loaded font.
 newtype Font
@@ -333,6 +338,11 @@ data Renderer = Renderer {
   -}
   setStrokeRadialGradient :: Point -> Double -> Double -> Color -> Color -> IO (),
   {-|
+  Sets a box gradient stroke with box area, corner radius, feather, inner and
+  outer Color
+  -}
+  setStrokeBoxGradient :: Rect -> Double -> Double -> Color -> Color -> IO (),
+  {-|
   Sets an image pattern stroke, with top given by Point, size of a single image
   given by size, rotation and alpha.
   -}
@@ -348,6 +358,11 @@ data Renderer = Renderer {
   inner and outer Color.
   -}
   setFillRadialGradient :: Point -> Double -> Double -> Color -> Color -> IO (),
+  {-|
+  Sets a box gradient fill with box area, corner radius, feather, inner and
+  outer Color
+  -}
+  setFillBoxGradient :: Rect -> Double -> Double -> Color -> Color -> IO (),
   {-|
   Sets an image pattern fill, with top given by Point, size of a single image
   given by size, rotation and alpha.